requirements.txt
opencv-python>=4.7.0
numpy>=1.24.0
gradio>=3.35.0
Pillow>=9.5.0
